requirements.txt
fastapi
uvicorn
apscheduler
sqlalchemy
requests
python-dateutil
psycopg2-binary
python-dotenv